refactor: remove engine::sat module
