+++ pumpkin-solver/src/engine/mod.rs
@@ -4,7 +4,7 @@ pub(crate) mod cp;
 mod debug_helper;
 pub(crate) mod nogoods;
 pub(crate) mod predicates;
-mod sat;
+mod restart_strategy;
 mod solver_statistics;
 pub(crate) mod termination;
 pub(crate) mod variables;
@@ -16,4 +16,5 @@ pub(crate) use cp::*;
 pub(crate) use debug_helper::DebugDyn;
 pub(crate) use debug_helper::DebugHelper;
 pub(crate) use domain_events::DomainEvents;
-pub use sat::*;
+pub use restart_strategy::RestartOptions;
+pub(crate) use restart_strategy::RestartStrategy;
