+++ pumpkin-solver/src/engine/restart_strategy.rs
