--- pumpkin-solver/src/engine/sat/mod.rs
@@ -1,4 +1,0 @@
-mod restart_strategy;
-
-pub use restart_strategy::RestartOptions;
-pub(crate) use restart_strategy::RestartStrategy;
